api_turismo_sustentable/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pandas==2.2.3
boto3==1.35.68
pyarrow==18.0.0
python-dotenv==1.0.1
google-genai==1.13.0
plotly==5.24.1
pydantic==2.10.2